+++ src/Helper/LocaleCodeHelper.php
@@ -1,0 +1,39 @@
+<?php declare(strict_types=1);
+
+namespace PaynlPayment\Shopware6\Helper;
+
+use PaynlPayment\Shopware6\Repository\Language\LanguageRepositoryInterface;
+use Shopware\Core\Framework\Context;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\System\Language\LanguageEntity;
+
+class LocaleCodeHelper
+{
+    /** @var LanguageRepositoryInterface */
+    private $languageRepository;
+
+    public function __construct(LanguageRepositoryInterface $languageRepository)
+    {
+        $this->languageRepository = $languageRepository;
+    }
+
+    public function getLocaleCodeFromContext(Context $context): string
+    {
+        $languageId = $context->getLanguageId();
+        $criteria = new Criteria([$languageId]);
+        $criteria->addAssociation('locale');
+        $criteria->setLimit(1);
+        /** @var LanguageEntity|null $language */
+        $language = $this->languageRepository->search($criteria, $context)->first();
+        if ($language === null) {
+            return 'en-GB';
+        }
+
+        $locale = $language->getLocale();
+        if (!$locale) {
+            return 'en-GB';
+        }
+
+        return $locale->getCode();
+    }
+}