+++ src/Repository/Order/OrderRepositoryInterface.php
@@ -7,6 +7,7 @@ use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\Event\EntityWrittenContainerEvent;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\EntitySearchResult;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\IdSearchResult;
 
 interface OrderRepositoryInterface
 {
@@ -28,6 +29,13 @@ interface OrderRepositoryInterface
      * @return EntitySearchResult
      */
     public function search(Criteria $criteria, Context $context): EntitySearchResult;
+
+    /**
+     * @param Criteria $criteria
+     * @param Context $context
+     * @return IdSearchResult
+     */
+    public function searchIds(Criteria $criteria, Context $context): IdSearchResult;
 
     /**
      * @param array<mixed> $data